+++ spec/fixtures/StoryboardLintTest/Pods/SomePod/Something.h
@@ -1,0 +1,2 @@
+@interface ClassFromCocoaPod : NSObject
+@end